The demo is only for tree construction
Because it is the only type of tree supported so far
